*Commit 4* - added on appearing and on disappearing events to state and country pages, added labels to state and country pages to display search results of entry string

+++ CoronaVirus/CountryData.cs
@@ -7,7 +7,7 @@ namespace CoronaVirus
     class CountryData
     {       
         public string country { get; set; }
-        public int updated { get; set; }
+        public long updated { get; set; }
         public int cases { get; set; }
         public int todayCases { get; set; }
         public int deaths { get; set; }
